ignore git directory in npm package
Thi should fix `EISGIT` errors during `npm install`. The error happened with npm v6.2.0 (probably other versions as well) when `npm i` was run multiple times in a repo (e.g. for package updates). It did not matter whether the FSPromise package was installed as a direct dependency of the repo or as a 2nd level dependency.

A very similar problem was documented here:
https://github.com/APSL/react-native-version-number/issues/13